Create ManualBreadcrumbs and use for task directory
